Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    float a;
-    scanf("%0.4f",&a);
+    double int a;
+    scanf("%ld",&a);
     printf("You entered: %ld",a);
     return 0;
 }
